fix issue #13

+++ FileHeader.py
@@ -3,7 +3,7 @@
 # @Author: lime
 # @Date:   2013-10-28 13:39:48
 # @Last Modified by:   lime
-# @Last Modified time: 2014-02-07 18:02:05
+# @Last Modified time: 2014-03-05 10:16:24
 
 import os
 import sys
@@ -17,6 +17,7 @@ import getpass
 import shutil
 import time
 import pickle
+import filecmp
 
 from datetime import datetime
 
@@ -27,6 +28,7 @@ else:
 
 
 PLUGIN_NAME = 'FileHeader'
+INSTALLED_PLUGIN_NAME = '%s.sublime-package' % PLUGIN_NAME
 PACKAGES_PATH = sublime.packages_path()
 PLUGIN_PATH = os.path.join(PACKAGES_PATH, PLUGIN_NAME)
 HEADER_PATH = os.path.join(PLUGIN_PATH, 'template/header')
@@ -37,16 +39,19 @@ IS_ST3 = sublime.version() >= '3'
 
 sys.path.insert(0, PLUGIN_PATH)
 
-
 def plugin_loaded():
     '''ST3'''
 
+    global LOADED
     global PACKAGES_PATH
     global PLUGIN_PATH
     global HEADER_PATH
     global BODY_PATH
     global INSTALLED_PLGIN_PATH
     global IS_ST3
+
+    PLUGIN_NAME = 'FileHeader'
+    INSTALLED_PLUGIN_NAME = '%s.sublime-package' % PLUGIN_NAME
 
     PACKAGES_PATH = sublime.packages_path()
     PLUGIN_PATH = os.path.join(PACKAGES_PATH, PLUGIN_NAME)
@@ -59,12 +64,16 @@ def plugin_loaded():
     sys.path.insert(0, PLUGIN_PATH)
 
     if INSTALLED_PLGIN_PATH != PLUGIN_PATH:
+        _ = os.path.join(PLUGIN_PATH, INSTALLED_PLUGIN_NAME)
+        if os.path.exists(_) and filecmp.cmp(_, INSTALLED_PLGIN_PATH):
+            return
+
         if os.path.exists(PLUGIN_PATH):
             try:
                 shutil.rmtree(PLUGIN_PATH)
             except:
                 pass
-    
+        
         if not os.path.exists(PLUGIN_PATH):
             os.mkdir(PLUGIN_PATH)
 
@@ -72,6 +81,8 @@ def plugin_loaded():
         for f in z.namelist():
             z.extract(f, PLUGIN_PATH)
         z.close()
+
+        shutil.copyfile(INSTALLED_PLGIN_PATH, _)
 
 
 def Window():
@@ -554,7 +565,7 @@ class FileHeaderListener(sublime_plugin.EventListener):
                 elif what == 'LAST_MODIFIED_TIME':                    
                     strings = datetime.now().strftime(get_strftime())
                 elif what == 'FILE_NAME':
-                    strings = 'undefined' if view.file_name() is None else view.file_name().split('/')[-1]
+                    strings = 'undefined' if view.file_name() is None else os.path.basename(view.file_name())
 
                 spaces = (index - space_start) * ' '
                 strings = spaces + strings
